# perf: run lint while constructing nodes