--- a/rust/parse_ast/src/lib.rs
+++ b/rust/parse_ast/src/lib.rs
@@ -1,6 +1,3 @@
-#![feature(ptr_internals)]
-use std::sync::Arc;
-
 use convert_ast::converter::AstConverter;
 use swc_common::sync::Lrc;
 use swc_common::{FileName, FilePathMapping, Globals, SourceMap, GLOBALS};
@@ -18,7 +15,7 @@ use error_emit::try_with_handler;
 mod error_emit;
 
 pub fn parse_ast(code: String, allow_return_outside_function: bool) -> Vec<u8> {
-  let cm = Arc::new(SourceMap::new(FilePathMapping::empty()));
+  let cm = Lrc::new(SourceMap::new(FilePathMapping::empty()));
   let target = EsVersion::EsNext;
   let syntax = Syntax::Es(EsConfig {
     allow_return_outside_function,
@@ -31,7 +28,7 @@ pub fn parse_ast(code: String, allow_return_outside_function: bool) -> Vec<u8> {
   let code_reference = Lrc::clone(&file.src);
   let comments = SequentialComments::default();
   GLOBALS.set(&Globals::default(), || {
-    let result = try_with_handler(&code_reference, &cm.clone(), target, |handler| {
+    let result = try_with_handler(&code_reference, |handler| {
       parse_js(
         cm,
         file,