--- a/src/Module.ts
+++ b/src/Module.ts
@@ -59,6 +59,7 @@ import {
 	error,
 	logAmbiguousExternalNamespaces,
 	logCircularReexport,
+	logDuplicateExportError,
 	logInconsistentImportAttributes,
 	logInvalidFormatForTopLevelAwait,
 	logInvalidSourcemapForError,
@@ -999,12 +1000,19 @@ export default class Module {
 		this.dynamicImports.push({ argument, id: null, node, resolution: null });
 	}
 
+	private hasExportNamedAs(name: string) {
+		return this.exports.has(name) || this.reexportDescriptions.has(name);
+	}
+
 	private addExport(
 		node: ExportAllDeclaration | ExportNamedDeclaration | ExportDefaultDeclaration
 	): void {
 		if (node instanceof ExportDefaultDeclaration) {
 			// export default foo;
 
+			if (this.hasExportNamedAs('default')) {
+				this.error(logDuplicateExportError('default'), node.start);
+			}
 			this.exports.set('default', {
 				identifier: node.variable.getAssignedVariableName(),
 				localName: 'default'
@@ -1016,6 +1024,9 @@ export default class Module {
 				// export * as name from './other'
 
 				const name = node.exported.name;
+				if (this.hasExportNamedAs(name)) {
+					this.error(logDuplicateExportError(name), node.exported.start);
+				}
 				this.reexportDescriptions.set(name, {
 					localName: '*',
 					module: null as never, // filled in later,
@@ -1034,6 +1045,9 @@ export default class Module {
 			this.addSource(source, node);
 			for (const { exported, local, start } of node.specifiers) {
 				const name = exported instanceof Literal ? exported.value : exported.name;
+				if (this.hasExportNamedAs(name)) {
+					this.error(logDuplicateExportError(name), start);
+				}
 				this.reexportDescriptions.set(name, {
 					localName: local instanceof Literal ? local.value : local.name,
 					module: null as never, // filled in later,
@@ -1049,6 +1063,9 @@ export default class Module {
 
 				for (const declarator of declaration.declarations) {
 					for (const localName of extractAssignedNames(declarator.id)) {
+						if (this.hasExportNamedAs(localName)) {
+							this.error(logDuplicateExportError(localName), declarator.id.start);
+						}
 						this.exports.set(localName, { identifier: null, localName });
 					}
 				}
@@ -1056,6 +1073,9 @@ export default class Module {
 				// export function foo () {}
 
 				const localName = (declaration.id as Identifier).name;
+				if (this.hasExportNamedAs(localName)) {
+					this.error(logDuplicateExportError(localName), declaration.id!.start);
+				}
 				this.exports.set(localName, { identifier: null, localName });
 			}
 		} else {
@@ -1065,6 +1085,9 @@ export default class Module {
 				// except for reexports, local must be an Identifier
 				const localName = (local as Identifier).name;
 				const exportedName = exported instanceof Identifier ? exported.name : exported.value;
+				if (this.hasExportNamedAs(exportedName)) {
+					this.error(logDuplicateExportError(exportedName), exported.start);
+				}
 				this.exports.set(exportedName, { identifier: null, localName });
 			}
 		}