--- a/src/ast/nodes/CatchClause.ts
+++ b/src/ast/nodes/CatchClause.ts
@@ -5,6 +5,7 @@ import type * as NodeType from './NodeType';
 import { UNKNOWN_EXPRESSION } from './shared/Expression';
 import { type GenericEsTreeNode, NodeBase } from './shared/Node';
 import type { PatternNode } from './shared/Pattern';
+import { VariableKind } from './shared/VariableKinds';
 
 export default class CatchClause extends NodeBase {
 	declare body: BlockStatement;
@@ -28,7 +29,7 @@ export default class CatchClause extends NodeBase {
 				this,
 				this.scope
 			);
-			this.param!.declare('parameter', UNKNOWN_EXPRESSION);
+			this.param!.declare(VariableKind.parameter, UNKNOWN_EXPRESSION);
 		}
 		super.parseNode(esTreeNode);
 	}