--- a/src/ast/nodes/Identifier.ts
+++ b/src/ast/nodes/Identifier.ts
@@ -2,6 +2,7 @@ import isReference, { type NodeWithFieldDefinition } from 'is-reference';
 import type MagicString from 'magic-string';
 import type { NormalizedTreeshakingOptions } from '../../rollup/types';
 import { BLANK } from '../../utils/blank';
+import { logRedeclarationError } from '../../utils/logs';
 import { PureFunctionKey } from '../../utils/pureFunctions';
 import type { NodeRenderOptions, RenderOptions } from '../../utils/renderHelpers';
 import type { DeoptimizableEntity } from '../DeoptimizableEntity';
@@ -13,7 +14,10 @@ import {
 	INTERACTION_CALLED,
 	NODE_INTERACTION_UNKNOWN_ACCESS
 } from '../NodeInteractions';
-import type FunctionScope from '../scopes/FunctionScope';
+import CatchScope from '../scopes/CatchScope';
+import ChildScope from '../scopes/ChildScope';
+import FunctionScope from '../scopes/FunctionScope';
+import type Scope from '../scopes/Scope';
 import { EMPTY_PATH, type ObjectPath, type PathTracker } from '../utils/PathTracker';
 import GlobalVariable from '../variables/GlobalVariable';
 import LocalVariable from '../variables/LocalVariable';
@@ -71,11 +75,44 @@ export default class Identifier extends NodeBase implements PatternNode {
 		}
 	}
 
+	private checkVarRedeclarationInScope(scope: Scope) {
+		if (!(scope instanceof CatchScope)) {
+			const declaredVariableKind = scope.variables.get(this.name)?.kind;
+			if (
+				declaredVariableKind === 'let' ||
+				declaredVariableKind === 'const' ||
+				declaredVariableKind === 'class' ||
+				(!(this.scope instanceof FunctionScope) && declaredVariableKind === 'function')
+			) {
+				this.scope.context.error(logRedeclarationError(this.name), this.start);
+			}
+		}
+		this.checkImportRedeclation(scope);
+	}
+
+	private checkImportRedeclation(scope: Scope) {
+		if (
+			scope === this.scope.context.module.scope &&
+			this.scope.context.module.importDescriptions.has(this.name)
+		) {
+			this.scope.context.error(logRedeclarationError(this.name), this.start);
+		}
+	}
+
 	declare(kind: string, init: ExpressionEntity): LocalVariable[] {
 		let variable: LocalVariable;
 		const { treeshake } = this.scope.context.options;
 		switch (kind) {
 			case 'var': {
+				const variableScope = this.scope.findLexicalBoundary();
+				let scope: Scope = this.scope;
+				while (true) {
+					this.checkVarRedeclarationInScope(scope);
+
+					if (scope === variableScope || !(scope instanceof ChildScope)) break;
+					scope = scope.parent;
+				}
+
 				variable = this.scope.addDeclaration(this, this.scope.context, init, true);
 				if (treeshake && treeshake.correctVarValueBeforeDeclaration) {
 					// Necessary to make sure the init is deoptimized. We cannot call deoptimizePath here.
@@ -84,6 +121,18 @@ export default class Identifier extends NodeBase implements PatternNode {
 				break;
 			}
 			case 'function': {
+				const declaredVariableKind = this.scope.variables.get(this.name)?.kind;
+				if (
+					declaredVariableKind === 'let' ||
+					declaredVariableKind === 'const' ||
+					declaredVariableKind === 'class' ||
+					(!(this.scope instanceof FunctionScope) &&
+						(declaredVariableKind === 'var' || declaredVariableKind === 'parameter'))
+				) {
+					this.scope.context.error(logRedeclarationError(this.name), this.start);
+				}
+				this.checkImportRedeclation(this.scope);
+
 				// in strict mode, functions are only hoisted within a scope but not across block scopes
 				variable = this.scope.addDeclaration(this, this.scope.context, init, false);
 				break;
@@ -91,10 +140,17 @@ export default class Identifier extends NodeBase implements PatternNode {
 			case 'let':
 			case 'const':
 			case 'class': {
+				if (this.scope.variables.has(this.name)) {
+					this.scope.context.error(logRedeclarationError(this.name), this.start);
+				}
+				this.checkImportRedeclation(this.scope);
+
 				variable = this.scope.addDeclaration(this, this.scope.context, init, false);
 				break;
 			}
 			case 'parameter': {
+				this.checkVarRedeclarationInScope(this.scope);
+
 				variable = (this.scope as FunctionScope).addParameterDeclaration(this);
 				break;
 			}