--- a/src/ast/nodes/shared/VariableKinds.ts
+++ b/src/ast/nodes/shared/VariableKinds.ts
@@ -1,0 +1,9 @@
+export const enum VariableKind {
+	// eslint-disable-next-line unicorn/prevent-abbreviations
+	var = 'var',
+	function = 'function',
+	let = 'let',
+	const = 'const',
+	class = 'class',
+	parameter = 'parameter'
+}