--- a/test/function/samples/ast-validations/duplicate-parameter-name/main.js
+++ b/test/function/samples/ast-validations/duplicate-parameter-name/main.js
@@ -1,0 +1,1 @@
+function foo(a,a) {}