… package.json
… rust/Cargo.lock
… rust/parse_ast/Cargo.toml
… src/ast/nodes/ObjectPattern.ts
… src/ast/nodes/Property.ts
… src/ast/nodes/RestElement.ts
… src/ast/nodes/VariableDeclaration.ts
… src/ast/nodes/VariableDeclarator.ts
… src/ast/nodes/shared/ClassNode.ts
… src/ast/nodes/shared/FunctionBase.ts
… src/ast/nodes/shared/FunctionNode.ts
… src/ast/nodes/shared/Pattern.ts
… src/ast/scopes/BlockScope.ts
… src/ast/scopes/CatchBodyScope.ts
… src/ast/scopes/CatchScope.ts
… src/ast/scopes/ChildScope.ts
… src/ast/scopes/ClassBodyScope.ts
… src/ast/scopes/FunctionScope.ts
… src/ast/scopes/ModuleScope.ts
… src/ast/scopes/ParameterScope.ts
… src/ast/scopes/Scope.ts
… src/ast/scopes/TrackingScope.ts
… src/ast/variables/ArgumentsVariable.ts
… src/ast/variables/ExportDefaultVariable.ts
… src/ast/variables/ExternalVariable.ts
… src/ast/variables/LocalVariable.ts
… src/ast/variables/NamespaceVariable.ts
… src/ast/variables/ParameterVariable.ts
… src/ast/variables/SyntheticNamedExportVariable.ts
… src/ast/variables/Variable.ts