<!--
  ⚡️ katchow! We ❤️ Pull Requests!

  If you remove or skip this template, you'll make the 🐼 sad and the mighty god
  of Github will appear and pile-drive the close button from a great height
  while making animal noises.

  Pull Request Requirements:
  * Please include tests to illustrate the problem this PR resolves.
  * Please lint your changes by running `npm run lint` before creating a PR.
  * Please update the documentation in `/docs` where necessary

  Please place an x (no spaces - [x]) in all [ ] that apply.
-->

This PR contains:

- [ ] bugfix
- [ ] feature
- [ ] refactor
- [ ] documentation
- [x] other

Are tests included?

- [ ] yes (_bugfixes and features will not be merged without tests_)
- [x] no

Breaking Changes?

- [ ] yes (_breaking changes will not be merged unless absolutely necessary_)
- [x] no

List any relevant issue numbers:

https://github.com/rollup/rollup/discussions/5205

<!--
If this PR resolves any issues, list them as

  resolves #1234

where 1234 is the issue number. This will help us with house-keeping as Github will automatically add a note to those issues stating that a potential fix exists. Once the PR is merged, Github will automatically close those issues.

If an issue is only solved partially or is relevant in some other way, just list the number without "resolves".
-->

### Description

- resolves #5244

This PR implements this idea (https://github.com/rollup/rollup/discussions/5205#discussioncomment-7288424) of running lint when initialising AST nodes instead of running in SWC side.

Running the benchmark in https://github.com/rollup/rollup/issues/5182 (actually I used the esbuild repo instead), I got ~300ms / ~5% improvement on my laptop.

<details>
<summary>Benchmark results</summary>

#### rollup 3.29.4
```
# BUILD: 6074ms, 933 MB / 943 MB
## initialize: 0ms, 3.55 kB / 9.81 MB
## generate module graph: 2951ms, 763 MB / 773 MB
- plugin 0 (stdin) - resolveId: 13ms, 2.94 MB / 771 MB
- plugin 0 (stdin) - load: 7ms, 827 kB / 772 MB
generate ast: 626ms, 211 MB / 773 MB
analyze ast: 1435ms, 741 MB / 773 MB
## sort and bind modules: 246ms, 38.5 MB / 812 MB
## mark included statements: 2877ms, 132 MB / 943 MB
treeshaking pass 1: 1029ms, 103 MB / 916 MB
treeshaking pass 2: 474ms, 24 MB / 940 MB
treeshaking pass 3: 178ms, -5.22 MB / 935 MB
treeshaking pass 4: 170ms, 6.52 MB / 942 MB
treeshaking pass 5: 205ms, 321 kB / 942 MB
treeshaking pass 6: 154ms, 14.3 MB / 956 MB
treeshaking pass 7: 136ms, -5.38 MB / 951 MB
treeshaking pass 8: 140ms, 3.45 MB / 954 MB
treeshaking pass 9: 129ms, 1.7 MB / 956 MB
treeshaking pass 10: 131ms, -13.8 MB / 942 MB
treeshaking pass 11: 124ms, 1.11 MB / 943 MB
# GENERATE: 571ms, 106 MB / 1.05 GB
## initialize render: 0ms, 3.46 kB / 945 MB
## generate chunks: 49ms, 3.67 MB / 949 MB
optimize chunks: 1ms, 580 kB / 950 MB
## render chunks: 503ms, 82.9 MB / 1.03 GB
## transform chunks: 18ms, 19.8 MB / 1.05 GB
## generate bundle: 0ms, 1.74 kB / 1.05 GB
# WRITE: 32ms, 19.5 MB / 1.07 GB
```

### rollup 4.1.4
```
# BUILD: 6398ms, 841 MB / 850 MB
## initialize: 1ms, -636 kB / 8.54 MB
## generate module graph: 3267ms, 663 MB / 672 MB
- plugin 0 (stdin) - resolveId: 14ms, 3.18 MB / 678 MB
- plugin 0 (stdin) - load: 8ms, 882 kB / 674 MB
generate ast: 1025ms, 136 MB / 678 MB
analyze ast: 1405ms, 663 MB / 678 MB
## sort and bind modules: 229ms, 46.4 MB / 718 MB
## mark included statements: 2901ms, 132 MB / 850 MB
treeshaking pass 1: 1060ms, 100 MB / 811 MB
treeshaking pass 2: 496ms, 36.6 MB / 847 MB
treeshaking pass 3: 177ms, -5.25 MB / 842 MB
treeshaking pass 4: 168ms, 6.51 MB / 849 MB
treeshaking pass 5: 195ms, -109 kB / 848 MB
treeshaking pass 6: 153ms, -1.96 MB / 846 MB
treeshaking pass 7: 143ms, 10.9 MB / 857 MB
treeshaking pass 8: 138ms, -12.7 MB / 845 MB
treeshaking pass 9: 117ms, 1.52 MB / 846 MB
treeshaking pass 10: 123ms, 2.79 MB / 849 MB
treeshaking pass 11: 119ms, 1.14 MB / 850 MB
# GENERATE: 565ms, 102 MB / 954 MB
## initialize render: 0ms, 3.46 kB / 852 MB
## generate chunks: 53ms, 2.49 MB / 855 MB
optimize chunks: 1ms, 493 kB / 856 MB
## render chunks: 494ms, 80.2 MB / 935 MB
## transform chunks: 18ms, 19 MB / 954 MB
## generate bundle: 0ms, 1.74 kB / 954 MB
# WRITE: 29ms, 19.9 MB / 974 MB
```

### This PR
```
# BUILD: 6072ms, 847 MB / 856 MB
## initialize: 0ms, 3.55 kB / 9.24 MB
## generate module graph: 2955ms, 669 MB / 678 MB
- plugin 0 (stdin) - resolveId: 13ms, 1.11 MB / 680 MB
- plugin 0 (stdin) - load: 7ms, 828 kB / 679 MB
generate ast: 710ms, 149 MB / 679 MB
analyze ast: 1385ms, 740 MB / 680 MB
## sort and bind modules: 234ms, 37 MB / 715 MB
## mark included statements: 2882ms, 142 MB / 856 MB
treeshaking pass 1: 1048ms, 99.6 MB / 816 MB
treeshaking pass 2: 489ms, 22 MB / 838 MB
treeshaking pass 3: 182ms, 10.2 MB / 848 MB
treeshaking pass 4: 173ms, 6.71 MB / 855 MB
treeshaking pass 5: 196ms, -509 kB / 855 MB
treeshaking pass 6: 157ms, -1.68 MB / 853 MB
treeshaking pass 7: 141ms, -5.45 MB / 847 MB
treeshaking pass 8: 133ms, 3.45 MB / 851 MB
treeshaking pass 9: 123ms, 1.47 MB / 852 MB
treeshaking pass 10: 115ms, 2.89 MB / 855 MB
treeshaking pass 11: 118ms, 1.2 MB / 856 MB
# GENERATE: 561ms, 101 MB / 959 MB
## initialize render: 0ms, 3.46 kB / 858 MB
## generate chunks: 50ms, 1.75 MB / 860 MB
optimize chunks: 1ms, 380 kB / 862 MB
## render chunks: 495ms, 73 MB / 933 MB
## transform chunks: 16ms, 26.2 MB / 959 MB
## generate bundle: 0ms, 1.74 kB / 959 MB
# WRITE: 32ms, 17.9 MB / 977 MB
```
</details>

<details>
<summary>Specs of my laptop</summary>

- CPU: Intel Core-i7 1360P
- Memory: DDR5-4800 32GB
- OS: Windows 11

</details>

### References
- https://github.com/acornjs/acorn/blob/520547b883a83f656bcb59ef196c1b24b867e41e/acorn/src/scope.js#L38-L69
- https://github.com/acornjs/acorn/blob/520547b883a83f656bcb59ef196c1b24b867e41e/acorn/src/statement.js#L926-L933
- https://github.com/rollup/rollup/blob/a6448b99f725d457e35821b73a865b5c4d4c6a61/src/ast/nodes/Identifier.ts#L110-L112