@sapphi-red 2023-10-23
Thanks for the review! I'm currently a bit busy, so I would take a look next month.
(Feel free to directly edit the PR, if you want to push this forward earlier.)