@lukastaegert 2023-11-08
I have some more I want to push tomorrow. Hopefully, I have it working by then. I added a ton of tests and even found some subtle bugs, and got a much better grasp on how catch scoping works...