@lukastaegert 2023-11-09
I fixed the remaining issues now. Now the logic is indeed distributed over the scopes, which feels more natural to me and avoids duplicating logic from the scoping.
I also refined the logic around catch scopes a bit. The problem is if a `var` in the body of a catch clause has the same name as the catch parameter, then we have both the hoisted var in the outer scope but the assignment is added to the parameter. For now I made them the same variable. This ensures that when variables are renamed due to conflicts, the renaming will be consistent.
This logic is still buggy, though, when you have a double nested catch scope, see this example: https://rollupjs.org/repl/?pr=5207&shareable=JTdCJTIyZXhhbXBsZSUyMiUzQW51bGwlMkMlMjJtb2R1bGVzJTIyJTNBJTVCJTdCJTIyY29kZSUyMiUzQSUyMmltcG9ydCUyMCcuJTJGZGVwLmpzJyU1Q24lNUNudHJ5JTIwJTdCJTVDbiUyMCUyMHRocm93JTIwbmV3JTIwRXJyb3IoJzEnKSUzQiU1Q24lN0QlMjBjYXRjaCUyMChlKSUyMCU3QiU1Q24lMjAlMjB0cnklMjAlN0IlNUNuJTIwJTIwJTIwJTIwdGhyb3clMjBuZXclMjBFcnJvcignMicpJTNCJTVDbiUyMCUyMCU3RCUyMGNhdGNoJTIwKGUpJTIwJTdCJTVDbiUyMCUyMCUyMCUyMHZhciUyMGUlMjAlM0QlMjAnZSclM0IlNUNuJTIwJTIwJTIwJTIwYXNzZXJ0LmVxdWFsKGUlMkMlMjAnZScpJTNCJTVDbiUyMCUyMCU3RCU1Q24lMjAlMjBhc3NlcnQuZXF1YWwoZS5tZXNzYWdlJTJDJTIwJzEnKSUzQiU1Q24lN0QlNUNuYXNzZXJ0LmVxdWFsKGUlMkMlMjB1bmRlZmluZWQpJTNCJTIyJTJDJTIyaXNFbnRyeSUyMiUzQXRydWUlMkMlMjJuYW1lJTIyJTNBJTIybWFpbi5qcyUyMiU3RCUyQyU3QiUyMmNvZGUlMjIlM0ElMjJ0cnklMjAlN0IlNUNuJTIwJTIwdGhyb3clMjBuZXclMjBFcnJvcignMScpJTNCJTVDbiU3RCUyMGNhdGNoJTIwKGUpJTIwJTdCJTVDbiUyMCUyMHRyeSUyMCU3QiU1Q24lMjAlMjAlMjAlMjB0aHJvdyUyMG5ldyUyMEVycm9yKCcyJyklM0IlNUNuJTIwJTIwJTdEJTIwY2F0Y2glMjAoZSklMjAlN0IlNUNuJTIwJTIwJTIwJTIwdmFyJTIwZSUyMCUzRCUyMCdlJyUzQiU1Q24lMjAlMjAlMjAlMjBhc3NlcnQuZXF1YWwoZSUyQyUyMCdlJyklM0IlNUNuJTIwJTIwJTdEJTVDbiUyMCUyMGFzc2VydC5lcXVhbChlLm1lc3NhZ2UlMkMlMjAnMScpJTNCJTVDbiU3RCU1Q25hc3NlcnQuZXF1YWwoZSUyQyUyMHVuZGVmaW5lZCklM0IlMjIlMkMlMjJpc0VudHJ5JTIyJTNBZmFsc2UlMkMlMjJuYW1lJTIyJTNBJTIyZGVwLmpzJTIyJTdEJTVEJTJDJTIyb3B0aW9ucyUyMiUzQSU3QiU3RCU3RA==

(Note that this is even more broken on current master...)
Output line 8 is asserting the wrong variable while line 12 will throw anyway because it is asserting a variable that does not even exist.
To that end, I wonder if maybe we should rewrite the declaration and make it two variables in this case. I.e.

```js
try {
  throw new Error('1');
} catch (e) {
  var e = 'e';
  assert.equal(e, 'e');
}
assert.equal(e, undefined);
```

becomes

```js
try {
  throw new Error('1');
} catch (e) {
  var e$1;
  e = 'e';
  assert.equal(e, 'e');
}
assert.equal(e$1, undefined);
```

This is indeed the correct logic (as weird as it looks) and might simplify things. Will need to think about this.