@lukastaegert 2023-11-10
Ok, it is more complicated than I first thought. So if the declaration uses destructuring and/or is part of a comma-separated declaration for multiple variables, there is some work I need to put in here. I found a simple hack: I make it two separate variables but "link" them so that including one includes the other and they are always rendered with the same name.
In the future if we find time, we might look into implementing a logic that always splits declaration and initialization for hoisted variables, which would also allow for some new minor tree-shaking opportunities.