@sapphi-red 2023-11-11
Thanks a lot! 💚 
I pushed the refactors suggested in the review.
I also checked some test cases in acorn and tried with the playground. There were some errors in some cases (https://github.com/rollup/rollup/issues/5244, https://github.com/swc-project/swc/issues/8268, https://github.com/swc-project/swc/issues/8269). But these errors happened with 4.3.0 too, so I think this PR isn't related directly 👍 .