@lukastaegert 2023-11-11
Great work, especially for checking the acorn test cases ❤️
So there is still some way to go with static checks, but I agree. I will have one last look if handling missing exports is a low-hanging fruit.