@lukastaegert 2023-11-11
I changed the handling of missing entry exports so that it shows a proper error instead of this bug. This error will also be shown for invalid cases like `export {default}`, but it is definitely better than before. It will not flag this as an error in non-entry modules unless the user tries to import that specific export. But at least no invalid code will be generated as otherwise, the export will just be removed. Still, I think at some point in the future we might want to flag invalid exports in files as syntax errors.